Add -services and address review comments
