+++ cmd/heartbeat/main.go
@@ -11,7 +11,7 @@ import (
 	"github.com/gorilla/websocket"
 	"github.com/m-lab/go/flagx"
 	"github.com/m-lab/go/rtx"
-	"github.com/m-lab/locate/cmd/heartbeat/messaging"
+	v2 "github.com/m-lab/locate/api/v2"
 	"github.com/m-lab/locate/connection"
 	"github.com/m-lab/locate/static"
 )
@@ -19,7 +19,9 @@ import (
 var (
 	heartbeatURL        string
 	hostname            string
+	experiment          string
 	registrationURL     = flagx.URL{}
+	services            = flagx.KeyValueArray{}
 	heartbeatPeriod     = static.HeartbeatPeriod
 	mainCtx, mainCancel = context.WithCancel(context.Background())
 )
@@ -28,7 +30,9 @@ func init() {
 	flag.StringVar(&heartbeatURL, "heartbeat-url", "ws://localhost:8080/v2/platform/heartbeat",
 		"URL for locate service")
 	flag.StringVar(&hostname, "hostname", "", "The service hostname")
+	flag.StringVar(&experiment, "experiment", "", "Experiment name")
 	flag.Var(&registrationURL, "registration-url", "URL for site registration")
+	flag.Var(&services, "services", "Maps experiment target names to their set of services")
 }
 
 func main() {
@@ -36,8 +40,11 @@ func main() {
 	rtx.Must(flagx.ArgsFromEnvWithLog(flag.CommandLine, false), "failed to read args from env")
 
 	// Load registration data.
-	r, err := messaging.LoadRegistration(mainCtx, hostname, registrationURL.URL)
+	r, err := LoadRegistration(mainCtx, hostname, registrationURL.URL)
 	rtx.Must(err, "could not load registration data")
+	// Populate flag values.
+	r.Experiment = experiment
+	r.Services = services.Get()
 	b, err := json.Marshal(r)
 	rtx.Must(err, "failed to marshal registration message, msg: %v", r)
 
@@ -61,7 +68,7 @@ func write(ws *connection.Conn) {
 		case <-mainCtx.Done():
 			return
 		case <-ticker.C:
-			healthMsg := messaging.Health{
+			healthMsg := v2.Health{
 				Hostname: hostname,
 				Score:    1.0,
 			}
