+++ cmd/heartbeat/testdata/registration.json
